fixed memory tests and tidied at the same time
