+++ cares_reinforcement_learning/algorithm/policy/MADDPG.py
@@ -105,7 +105,7 @@ class MADDPG(MARLAlgorithm[dict[str, np.ndarray]]):
 
             agent_observation = SARLObservation(
                 vector_state=obs_i,
-                avail_actions=avail_i,
+                available_actions=avail_i,
             )
 
             agent_sample = agent.act(agent_observation, evaluation)
