+++ cares_reinforcement_learning/algorithm/policy/MAPPO.py
@@ -129,7 +129,7 @@ class MAPPO(MARLAlgorithm[dict[str, np.ndarray]]):
 
             agent_observation = SARLObservation(
                 vector_state=obs_i,
-                avail_actions=avail_i,
+                available_actions=avail_i,
             )
 
             agent_sample = agent.act(
