+++ cares_reinforcement_learning/algorithm/policy/MASAC.py
@@ -92,7 +92,7 @@ class MASAC(MARLAlgorithm[dict[str, np.ndarray]]):
 
             agent_observation = SARLObservation(
                 vector_state=obs_i,
-                avail_actions=avail_i,
+                available_actions=avail_i,
             )
 
             agent_sample = agent.act(agent_observation, evaluation)
